+++ compose-miuix/shared/src/commonMain/kotlin/com/perol/pixez/shared/ui/navigation/RootContent.kt
@@ -286,6 +286,7 @@ fun RootContent(
                         downloadRepository = downloadRepository,
                         banRepository = banRepository,
                         historyRepository = historyRepository,
+                        onIllustClick = component::onIllustClicked,
                     )
 
                     is Child.UserDetail -> UserDetailScreen(